adk_agents/player-6/requirements.txt
google-adk>=0.1.0
google-cloud-aiplatform
google-auth
google-generativeai>=0.3.0
python-dotenv>=1.0.0 